[ci] test: consolidate gather backward regressions
