+++ .agents/knowledge/constraints.md
@@ -219,5 +219,5 @@ Core files:
 
 28. **In-place collective reductions in backward must own their gradient buffer**
     - Autograd can pass the same incoming gradient to multiple branches. `.contiguous()` does not copy an already contiguous tensor, so reducing that tensor in place can silently change a sibling branch's gradient and the caller's `grad_outputs`.
-    - `_Gather.backward` clones into contiguous storage before its optional sum reduction, as `_GatherConcatSP.backward` also does. Preserve the separate sum/no-sum and scale/no-scale semantics; regression tests live in `tests/parallel/ulysses/test_gather_backward.py` for real Gloo and NCCL collectives.
+    - `_Gather.backward` clones into contiguous storage before its optional sum reduction, as `_GatherConcatSP.backward` also does. Preserve the separate sum/no-sum and scale/no-scale semantics; regression tests live in `tests/parallel/ulysses/test_all_gather.py` for real Gloo and NCCL collectives.
     - A regression's reference collective must also use a contiguous buffer for NCCL. Make only the reference clone contiguous; preserve the layout of the actual incoming gradient so transposed, narrowed and expanded inputs remain covered.
